Rename and refactor the details form
This reanmes the GcseQualificationDetailsForm to
MathsGcseGradeForm as this is now the role responsbility
of the form.

I've stripped out all of the logic to do with award year
and generally just cleaned it up.

This also removes all the logic to do with awar year.
