+++ spec/forms/candidate_interface/restructured_work_history/work_history_break_form_spec.rb
@@ -22,7 +22,6 @@ RSpec.describe CandidateInterface::RestructuredWorkHistory::WorkHistoryBreakForm
   describe 'validations' do
     it { is_expected.to validate_presence_of(:reason) }
 
-
     okay_text = Faker::Lorem.sentence(word_count: 400)
     long_text = Faker::Lorem.sentence(word_count: 401)
